fixup! Move GroupByKey expansion into DirectPipelineRunner
